Cost Artemether/Lumefantrine treatment

Note the cost per dose is for a single dose (20/120 mg). A treatment course typically constitutes Artemether + lumefantrine given twice a day for 3 days following weight-based guidelines:

5 to <15 kg: 20/120 mg

15 to <25 kg: 40/240 mg

35 to <35 kg: 60/360 mg

>= 35 kg: 80/480 mg

So course for a single adult (weighing >=35kg) may constitute 3 days x 2 times daily x 4 doses (4 x 20/120mg = 80/480mg) = 24 doses.

Usage

cost_al(n_doses, cost_per_dose = 0.3)

Arguments

n_doses

Number of tests

cost_per_dose

Cost per dose is for a single dose (20/120 mg)

Value

AL costs

References

dosing information

https://www.who.int/publications/i/item/guidelines-for-malaria

cost_per_dose

Current default is the average cost for 20/120mg.

The Global Fund Pooled Procurement Mechanism Reference Pricing: Antimalarial medicines, version: quarter 1, 2022

https://www.theglobalfund.org/en/sourcing-management/health-products/antimalarial-medicines/.